Split appointment list item into its own component.
